requirements.txt
requests
pymongo
transformers
python-dotenv